Move mesh.cf code into fileformats.cf, and tests likewise
Fix imports in cf ugrid tests.
