+++ lib/iris/mesh/__init__.py
@@ -9,8 +9,8 @@ Based on CF UGRID Conventions (v1.0), https://ugrid-conventions.github.io/ugrid-
 """
 
 from iris.fileformats.netcdf.saver import save_mesh
-from iris.fileformats.netcdf.ugrid_load import load_mesh, load_meshes
 
+# from iris.fileformats.netcdf.ugrid_load import load_mesh, load_meshes
 from ..config import get_logger
 from .components import Connectivity, MeshCoord, MeshXY
 from .utils import recombine_submeshes
@@ -19,8 +19,6 @@ __all__ = [
     "Connectivity",
     "MeshCoord",
     "MeshXY",
-    "load_mesh",
-    "load_meshes",
     "recombine_submeshes",
     "save_mesh",
 ]
