--- lib/iris/tests/unit/mesh/cf/__init__.py
@@ -1,5 +1,0 @@
-# Copyright Iris contributors
-#
-# This file is part of Iris and is released under the BSD license.
-# See LICENSE in the root of the repository for full licensing details.
-"""Unit tests for the :mod:`iris.mesh.cf` package."""
